Move Domain to org.h2.schema and remove incorrect information from documentation
