+++ h2/src/main/org/h2/command/ddl/CreateDomain.java
@@ -8,9 +8,9 @@ package org.h2.command.ddl;
 import java.util.ArrayList;
 import org.h2.api.ErrorCode;
 import org.h2.command.CommandInterface;
-import org.h2.engine.Domain;
 import org.h2.engine.Session;
 import org.h2.message.DbException;
+import org.h2.schema.Domain;
 import org.h2.schema.Schema;
 import org.h2.table.Column;
 import org.h2.table.Table;
