+++ h2/src/main/org/h2/command/ddl/DropDomain.java
@@ -11,10 +11,10 @@ import org.h2.command.CommandInterface;
 import org.h2.constraint.ConstraintActionType;
 import org.h2.constraint.ConstraintDomain;
 import org.h2.engine.Database;
-import org.h2.engine.Domain;
 import org.h2.engine.Session;
 import org.h2.expression.Expression;
 import org.h2.message.DbException;
+import org.h2.schema.Domain;
 import org.h2.schema.Schema;
 import org.h2.table.Column;
 import org.h2.table.Table;
