+++ h2/src/main/org/h2/constraint/ConstraintDomain.java
@@ -9,13 +9,13 @@ import java.util.HashSet;
 
 import org.h2.api.ErrorCode;
 import org.h2.command.Parser;
-import org.h2.engine.Domain;
 import org.h2.engine.Session;
 import org.h2.expression.Expression;
 import org.h2.expression.ExpressionVisitor;
 import org.h2.index.Index;
 import org.h2.message.DbException;
 import org.h2.result.Row;
+import org.h2.schema.Domain;
 import org.h2.schema.Schema;
 import org.h2.table.Column;
 import org.h2.table.Table;
