+++ h2/src/main/org/h2/schema/Schema.java
@@ -19,7 +19,6 @@ import org.h2.engine.Database;
 import org.h2.engine.DbObject;
 import org.h2.engine.DbObjectBase;
 import org.h2.engine.DbSettings;
-import org.h2.engine.Domain;
 import org.h2.engine.FunctionAlias;
 import org.h2.engine.Right;
 import org.h2.engine.Session;
